ems-dj/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
decorator==4.4.2
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
idna==3.3
imageio==2.16.1
imageio-ffmpeg==0.4.5
install==1.3.5
moviepy==1.0.3
mysqlclient==2.1.0
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.0.0
proglog==0.1.9
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
tqdm==4.63.1
urllib3==1.26.9
